+++ lib/govspeak.rb
@@ -354,6 +354,14 @@ module Govspeak
       render_image(ImagePresenter.new(image))
     end
 
+    # This is an alternative syntax for embedding attachments using a readable id (expected
+    # to be a unique variation of a filename). This syntax is being used by
+    # Content Publisher and should be considered experimental as it is likely
+    # to be iterated in the short term.
+    extension('Attachment', /#{NEW_PARAGRAPH_LOOKBEHIND}\[Attachment:\s*(.*?)\s*\]/) do |attachment_id|
+      %{<govspeak-embed-attachment id="#{attachment_id}"></govspeak-embed-attachment>}
+    end
+
   private
 
     def kramdown_doc